% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_ana.R
\name{anova_ana}
\alias{anova_ana}
\title{Performs anova analysis
Performs an anova across all response variables. The function takes 2
independent variables.}
\usage{
anova_ana(
  metabolomics_data,
  met_col,
  analysis_type,
  sample_col,
  metadata,
  normalization,
  factor_col1,
  factor_col2,
  p_adjust
)
}
\arguments{
\item{metabolomics_data}{metabolomics data}

\item{met_col}{column with metabolite names}

\item{analysis_type}{type of analysis ex-GCMS, HILIAC positive ion mode.}

\item{sample_col}{the column name having samples}

\item{metadata}{Metadata}

\item{normalization}{normalization methods any of the three: half_of_min, remove_NAs and 50percent
\enumerate{
\item half_of_min: where the NAs are replaced by half of min values in the data
\item remove_NAs: where Cols with NAs values are removed
\item 50percent: where cols with more than 50\% NAs values are removed
}}

\item{factor_col1}{column name of first independent variable}

\item{factor_col2}{column name of second independent variable}

\item{p_adjust}{Method for p value adjustment, i.e. "fdr"}
}
\description{
Performs anova analysis
Performs an anova across all response variables. The function takes 2
independent variables.
}
\examples{
anova_result = anova_ana(metabolomics_data=data,met_col="metabolite_name",analysis_type=c('Phospholipids, Chol. esters and Diacylglycerols','Sphingolipids'), metadata=metadata,normalization="50percent", factor_col1="TreatmentGroup",factor_col2="SamplingTimePoint",p_adjust="fdr")
}
