% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_linkedto_pathways.R
\name{comp_linkedto_pathways}
\alias{comp_linkedto_pathways}
\title{counts total no of compounds that are linked to KEGG pathways}
\usage{
comp_linkedto_pathways()
}
\description{
counts total no of compounds that are linked to KEGG pathways
}
\examples{
ls_path =comp_linkedto_pathways()
}
