% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundinfo.R
\name{compoundinfo}
\alias{compoundinfo}
\title{Counts total no of compound in a pathway that has been linked to significant metabolites
This will help in calculating HG score for metabolite enrichment}
\usage{
compoundinfo(met_path, sps)
}
\arguments{
\item{met_path}{dataframe of pathways mapped to significant metabolites as obtained from met_pathways}

\item{sps}{species name}
}
\description{
Counts total no of compound in a pathway that has been linked to significant metabolites
This will help in calculating HG score for metabolite enrichment
}
\examples{
compound_count = compoundinfo(met_path, sps)
}
