% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_refmet.R
\name{convert_refmet}
\alias{convert_refmet}
\title{Associates refmet class to the metabolites}
\usage{
convert_refmet(metdata)
}
\arguments{
\item{metdata}{metabolomics data retrieved or uploaded by the user}
}
\description{
Associates refmet class to the metabolites
}
\examples{
refmet_class= convert_refmet(metdata)
}
