% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme_gene_info.R
\name{enzyme_gene_info}
\alias{enzyme_gene_info}
\title{Gets gene and enzyme info}
\usage{
enzyme_gene_info(df_metenrichment, sps, classm)
}
\arguments{
\item{df_metenrichment}{metenrichment score dataframe having metabolites, metabolite class and their enrichment score obtained
from metclassenrichment}

\item{sps}{species name}

\item{classm}{class name of the metabolite whether sub_class, main_class or super_class}
}
\description{
Gets gene and enzyme info
}
\examples{
met_gene_info = enzyme_gene_info (metenrichment, "hsa","sub_class")
}
