% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmwstudies.R
\name{getmwstudies}
\alias{getmwstudies}
\title{Gets metabolomics data, metadata and metabolite info from Metabolomics Workbench using REST service}
\usage{
getmwstudies(studyid, typedata)
}
\arguments{
\item{studyid}{from MW and typedata whether metadata (factors) or metabolic data}

\item{typedata}{can be "factors" for metadata, "data" for metabolomics data and 'metabolites' for metabolite info}
}
\description{
Gets metabolomics data, metadata and metabolite info from Metabolomics Workbench using REST service
}
\examples{
data = getmwstudies('ST001304', 'data')
}
