% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ls_path}
\alias{ls_path}
\title{compound linked to kegg pathways}
\format{
A list with 6148 elements:
}
\source{
ls_path = comp_linkedto_pathways()
}
\usage{
data(ls_path)
}
\description{
A dataset showing count of kegg compounds that are linked to kegg pathways
}
\examples{
data(ls_path)
}
\keyword{datasets}
