% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_keggid.R
\name{map_keggid}
\alias{map_keggid}
\title{Assigns kegg id to the significant metabolites}
\usage{
map_keggid(sig_metabolites)
}
\arguments{
\item{sig_metabolites}{dataframe having significant metabolite with metabolite class}
}
\description{
Assigns kegg id to the significant metabolites
}
\examples{
sig_metabolites_kegg_id= map_keggid(sig_metabolites)
}
