% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapspspath.R
\name{mapspspath}
\alias{mapspspath}
\title{Maps species specific pathways
This is important because sometimes some species may not have some pathways}
\usage{
mapspspath(sps, rxn_pathway_df)
}
\arguments{
\item{sps}{species name}

\item{rxn_pathway_df}{dataframe having pathway information on metabolites as obtained from pathinfo}
}
\description{
Maps species specific pathways
This is important because sometimes some species may not have some pathways
}
\examples{
org_pathways_map= mapspspath(sps=sps,rxn_pathway_df= res_path)
}
