% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_enrichmentscore.R
\name{path_enrichmentscore}
\alias{path_enrichmentscore}
\title{Calculates pathway enrichment score of the pathways linked to significant metabolites}
\usage{
path_enrichmentscore(
  met_path,
  sig_metabolite_kegg_id,
  ls_path,
  refmet_class,
  sps,
  padj,
  kegg_comp_path = FALSE
)
}
\arguments{
\item{met_path}{dataframe of pathways mapped to significant metabolites as obtained from met_pathways}

\item{sig_metabolite_kegg_id}{dataframe of significantly altered metabolites for HG calculation}

\item{ls_path}{list of all the KEGG compounds linked to pathways. This can be obtained by running comp_linkedto_pathways().R or is also available
please use ls_path when you want to calculate HG score depending on all the compounds in KEGG linked to a pathway
sysdata.rda}

\item{refmet_class}{dataframe of metabolite data with refmet class for a study, alternatively,
data obtained from rest can also be given here which is yet not associated to refmet class
We need this to get the number of all the metabolites detected in the study for calculating HG score.}

\item{sps}{species name}

\item{padj}{Method for p value adjustment, i.e. "fdr"}

\item{kegg_comp_path}{TRUE or FALSE depending you want to use the ls_path as your background. Default is FALSE, where background set
is the total metabolite in a study}
}
\description{
Calculates pathway enrichment score of the pathways linked to significant metabolites
}
\examples{
kegg_es = path_enrichmentscore(met_path,sig_metabolite_kegg_id, ls_path,refmet_class,sps='hsa',padj='fdr')
}
