% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathinfo.R
\name{pathinfo}
\alias{pathinfo}
\title{Provides pathway or orthology information from reaction data of significant metabolite}
\usage{
pathinfo(res, col_name)
}
\arguments{
\item{res}{dataframe having metabolite information along with reaction information as obtained from rxninfo}

\item{col_name}{"PATHWAY" or "ORTHOLOGY"}
}
\description{
Provides pathway or orthology information from reaction data of significant metabolite
}
\examples{
res_path= pathinfo(res, 'PATHWAY') ###rxn pathway
}
