% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plots heatmap}
\usage{
plot_heatmap(met_path, shorten_name = TRUE, refmet_name = FALSE)
}
\arguments{
\item{met_path}{dataframe of pathways mapped to significant metabolites}

\item{shorten_name}{default TRUE if metabolite names are too big}

\item{refmet_name}{default false, set to true if you want to plot according to refmet names}
}
\description{
Plots heatmap
}
\examples{
plot_heatmap(met_path)
}
