% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_met_enrichment.R
\name{plot_met_enrichment}
\alias{plot_met_enrichment}
\title{Plots metabolite enrichment score}
\usage{
plot_met_enrichment(df_metenrichment, metclass, enrich_stats, no)
}
\arguments{
\item{df_metenrichment}{metenrichment score dataframe having metabolites, metabolite class and their enrichment score obtained
from metclassenrichment}

\item{metclass}{= sub_class, main_class or super_class}

\item{enrich_stats}{HG for hypergeometric score ## leaves room for further including of KS stats}

\item{no}{number of significant metabolites that should be in a class, default 1, should increase to 3 or more for a
better statistically significant score. Default is chosen as 1 so as to get information on all the significant metabolites without any filtering.}
}
\description{
Plots metabolite enrichment score
}
\examples{
plot_met_enrichment(metenrichment, "sub_class","HG", no=1)
}
