% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_networks.R
\name{plot_pathway_networks}
\alias{plot_pathway_networks}
\title{Plots 1st degree pathway network of only pathways having significant pathway score}
\usage{
plot_pathway_networks(met_path, kegg_es, colorEdge)
}
\arguments{
\item{met_path}{dataframe of pathways mapped to significant metabolites}

\item{kegg_es}{dataframe of enrichment score of pathways}

\item{colorEdge}{default TRUE to color the edges by pathways}
}
\description{
Plots 1st degree pathway network of only pathways having significant pathway score
}
\examples{
plot_pathway_networks (met_path,kegg_es, TRUE)
}
