% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plots Volcano plot}
\usage{
plot_volcano(
  stats_metabolites,
  thres_pval,
  thres_log2foldchange,
  text_nooverlap
)
}
\arguments{
\item{stats_metabolites}{dataframe obtained by running sig_metabolites.R}

\item{thres_pval}{p value threshold}

\item{thres_log2foldchange}{log2fold change threshold}

\item{text_nooverlap}{TRUE to avoid text overlap, else FALSE}
}
\description{
Plots Volcano plot
}
\examples{
plot_volcano(stats_metabolites, thres_pval= 0.05,thres_log2foldchange = 0.5, text_nooverlap = TRUE)
}
