% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_substrate.R
\name{react_substrate}
\alias{react_substrate}
\title{Gets information whether metabolite is a product or substrate}
\usage{
react_substrate(met_gene_info)
}
\arguments{
\item{met_gene_info}{dataframe with gene information obtained from enzyme_gene_info}
}
\description{
Gets information whether metabolite is a product or substrate
}
\examples{
rclass_info = react_substrate(met_gene_info)
}
