% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxninfo.R
\name{rxninfo}
\alias{rxninfo}
\title{Gives details of reactions for the metabolites}
\usage{
rxninfo(class_mapped_data)
}
\arguments{
\item{class_mapped_data}{metenrichment score dataframe having metabolites, metabolite class and their enrichment score obtained
from metclassenrichment}
}
\description{
Gives details of reactions for the metabolites
}
\examples{
res= rxninfo(df_metenrichment)
}
