% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metcountplot.R
\name{metcountplot}
\alias{metcountplot}
\title{counts and plots significant metabolite}
\usage{
metcountplot(df_metclass, metclass, plotting, thres_logfC)
}
\arguments{
\item{df_metclass}{dataframe having significant metabolite with metabolite class as obtained from map_refmet_class.R}

\item{metclass}{sub_class, main_class or super_class}

\item{thres_logfC}{log2 fold change threshold}

\item{plot}{logical TRUE or FALSE, default true}
}
\description{
counts and plots significant metabolite
}
\examples{
count_changes = metcountplot(df_metclass=sig_metabolites_pubchem, metclass='sub_class', plotting=TRUE, thres_logfC = 0.5)
}
